+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/data/data-sql-view.js
@@ -1,6 +1,7 @@
 var CoreView = require('backbone/core-view');
 var CodeMirrorView = require('../../../../components/code-mirror/code-mirror-view');
 var ErrorTemplate = require('./code-mirror-error.tpl');
+var FactoryHints = require('../../../editor-hints/factory-hints');
 var SQLHints = require('../../../editor-hints/sql-hints');
 
 module.exports = CoreView.extend({
@@ -19,9 +20,10 @@ module.exports = CoreView.extend({
 
     this._initBinds();
 
-    SQLHints.init({
+    FactoryHints.init({
       querySchemaModel: this._querySchemaModel,
-      layerDefinitionModel: this._layerDefinitionModel
+      layerDefinitionModel: this._layerDefinitionModel,
+      tokens: SQLHints
     });
   },
 
@@ -44,7 +46,7 @@ module.exports = CoreView.extend({
   },
 
   _initViews: function () {
-    var hints = SQLHints.reset().hints;
+    var hints = FactoryHints.reset().hints;
     this.codeMirrorView = new CodeMirrorView({
       model: this._codemirrorModel,
       hints: hints,
