+++ lib/assets/javascripts/cartodb3/editor/style/style-cartocss-view.js
@@ -1,5 +1,6 @@
 var CoreView = require('backbone/core-view');
 var CodeMirrorView = require('../../components/code-mirror/code-mirror-view');
+var FactoryHints = require('../editor-hints/factory-hints');
 var CSSHints = require('../editor-hints/css-hints');
 
 module.exports = CoreView.extend({
@@ -22,9 +23,10 @@ module.exports = CoreView.extend({
 
     this._initBinds();
 
-    CSSHints.init({
+    FactoryHints.init({
       querySchemaModel: this._querySchemaModel,
-      layerDefinitionModel: this._layerDefinitionModel
+      layerDefinitionModel: this._layerDefinitionModel,
+      tokens: CSSHints
     });
   },
 
@@ -41,7 +43,7 @@ module.exports = CoreView.extend({
   },
 
   _initViews: function () {
-    var hints = CSSHints.reset().hints;
+    var hints = FactoryHints.reset().hints;
     this.codeMirrorView = new CodeMirrorView({
       model: this._codemirrorModel,
       hints: hints,
